Move more graphs
