+++ grakn-test-tools/src/main/java/ai/grakn/test/graphs/GenealogyGraph.java
@@ -16,13 +16,17 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.graphs;
+package ai.grakn.test.graphs;
 
 import ai.grakn.GraknGraph;
-import ai.grakn.test.graphs.TestGraph;
 
 import java.util.function.Consumer;
 
+/**
+ *
+ * @author Kasper Piskorski
+ *
+ */
 public class GenealogyGraph extends TestGraph {
 
     final private static String ontologyFile = "genealogy/ontology.gql";
