+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/ExplanationTest.java
@@ -20,7 +20,7 @@ package ai.grakn.test.graql.reasoner;
 
 import ai.grakn.GraknGraph;
 import ai.grakn.concept.Concept;
-import ai.grakn.graphs.GenealogyGraph;
+import ai.grakn.test.graphs.GenealogyGraph;
 import ai.grakn.graphs.GeoGraph;
 import ai.grakn.graql.Graql;
 import ai.grakn.graql.MatchQuery;
