+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/inference/GenealogyTest.java
@@ -19,7 +19,7 @@
 package ai.grakn.test.graql.reasoner.inference;
 
 import ai.grakn.concept.Concept;
-import ai.grakn.graphs.GenealogyGraph;
+import ai.grakn.test.graphs.GenealogyGraph;
 import ai.grakn.graql.Graql;
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.QueryBuilder;
